moved scripts to seperate file
